refactor fixture definition so it is more widely accessible
